% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bathy.R
\name{add_bathy}
\alias{add_bathy}
\title{Add bathymetry data, given locations}
\usage{
add_bathy(x, lat_var, lon_var, z_radius = 2.5, bathy_path)
}
\arguments{
\item{x}{A data frame with location data to associate with bathymetry data}

\item{lat_var}{name of variable in \code{x} with latitude data}

\item{lon_var}{name of variable in \code{x} with longitude data}

\item{z_radius}{Distance in km around each location that seafloor depth and slope will be calculated. This can be a single value or a vector of different values the same length as the number of rows of the data input. Default: 2.5 km}

\item{bathy_path}{A directory path to the folder containing all bathymetry data. File names in the folder must be like this :'NEPACseafloor-116.000-30.000-115.000-31.000.csv'. If this input is NULL, data will be pulled from online, thus requiring internet access. (Note that there may be some access and speed issues with pulling the data from the web.)}
}
\value{
A dataframe with associated bathymetry data as new columns in the dataset
}
\description{
Pulls bathymetry data from either the internet or a provided database and associates it with tag locations. Note: code obtained from Eric Keene (NOAA) via David Sweeney (MarEcoTel) with minor modification. To use, you will need bathymetry data. You can download it from the Drive FreakinBeakinTagData > Acoustic Preliminary Results > Resources-Seafloor.zip or by using \code{\link[FBtagtools]{download_bathymetry}} (which just locates and downloads that zip file for you).
}
\examples{
#examples not yet provided, sorry :(
}
