% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_event_depths.R
\name{add_event_depths}
\alias{add_event_depths}
\title{Add time columns to time-series}
\usage{
add_event_depths(x, start_x, z, sampling_rate, start_offset = 0, tagon_time)
}
\arguments{
\item{x}{input data: vector or data frame with event times as datetimes, or in seconds since tagon. Datetimes are assumed to be UTC.}

\item{start_x}{name (no quotes needed) of variable in \code{x} that gives event times. Not needed if \code{x} is a vector or has only one column.}

\item{z}{depth data: sensor data structure, vector, matrix, tibble, or data frame. If not a sensor structure, then \code{sampling_rate} is required.}

\item{sampling_rate}{sampling rate of depth in Hz (required if x is not a sensor data structure, and ignored if it is)}

\item{start_offset}{time (in seconds) between start of tag recording and the first recorded sample in \code{z}. Ignored (and pulled from metadata) if \code{x} is a sensor data structure. default: 0}

\item{tagon_time}{string giving the tag recording start time. Can be found in tag_dataset$info$dephist_device_datetime_start. Format: day month year hours minutes seconds. Required if times in \code{x} are datetimes.}
}
\value{
A data.frame with the data (one row per event) and additional column depth (in meters) of the animal at the start time of the event. Note: if you want to grab and summarize depth data for longer intervals, consider \code{\link[FBtagtools]{interval_join}} or \code{\link[tagtools]{find_dives}} and \code{\link[tagtools]{dive_stats}}
}
\description{
Add time (in seconds since tagon and perhaps as datetimes in UTC and/or local timezone) to a tibble or data-frame with regularly sampled data
}
\examples{
Examples will go here
}
