% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_event_times.R
\name{add_event_times}
\alias{add_event_times}
\title{Add times of events (UTC or seconds_since)}
\usage{
add_event_times(x, start_x, tagon_time)
}
\arguments{
\item{x}{input data frame: vector, tibble, or data frame.  There should be one column that gives either seconds_since_tagon or UTC time already present.}

\item{start_x}{name of the variable in x that gives the times}

\item{tagon_time}{string giving the tag recording start time (or the \code{info} sensor data structure). Can be found in tag_dataset$info$dephist_device_datetime_start. Format: day month year hours minutes seconds.}
}
\description{
Add time (in seconds since tagon or datetimes in UTC and/or local timezone) to a tibble or data-frame with event times (conversion between datetime and seconds_since_tagon formats)
}
\examples{
Examples will go here
}
