% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_interval_rls.R
\name{add_interval_rls}
\alias{add_interval_rls}
\title{Add RL information for time intervals}
\usage{
add_interval_rls(x, ping_data, start_x, end_x, start_ping)
}
\arguments{
\item{x}{data frame or tibble with each row giving data about a time interval (for one tag)}

\item{ping_data}{data frame or tibble for one tag, with each row giving rl data from one sound (for example, from \code{\link[FBtagtools]{extract_rls}})}

\item{start_x}{name of variable in x (quotes not needed) with start times of intervals, in seconds since start of tag recording}

\item{end_x}{name of variable in x (quotes not needed) with end times of intervals in seconds since start of tag recording}

\item{start_ping}{name of variable in ping_data giving start times of pings in seconds since start of tag recording}
}
\value{
x with additional columns giving average, min, max, etc. RLs in each time interval
}
\description{
Add information about broadband RMS RLs to a dataset where each row contains data on a time interval, with start/end times of the interval given in seconds since start of tag record
}
