% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sensor_times.R
\name{add_sensor_times}
\alias{add_sensor_times}
\title{Add time columns to sensor data time-series}
\usage{
add_sensor_times(
  x,
  sampling_rate,
  start_offset = 0,
  add_utc_time = FALSE,
  tagon_time
)
}
\arguments{
\item{x}{input data: sensor data structure, vector, matrix, tibble, or data frame.  If \code{x} is not a tag sensor data structure, then \code{sampling_rate} is required input.}

\item{sampling_rate}{sampling rate of x in Hz (required if x is not a sensor data structure, and ignored if it is)}

\item{start_offset}{time (in seconds) between start of tag recording and the first recorded sample. Ignored (and pulled from metadata) if \code{x} is a sensor data structure. default: 0}

\item{add_utc_time}{Logical: should UTC time column be added? Default is FALSE.}

\item{tagon_time}{string giving the tag recording start time. Can be found in tag_dataset$info$dephist_device_datetime_start. Format: day month year hours minutes seconds. Required if \code{add_dttm} is TRUE.}
}
\description{
Add time (in seconds since tagon and perhaps as datetimes in UTC and/or local timezone) to a tibble or data-frame with regularly sampled data
}
\examples{
Examples will go here
}
