% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exposure.R
\name{check_exposure}
\alias{check_exposure}
\title{Add time columns to time-series}
\usage{
check_exposure(x, start_x, rl_data)
}
\arguments{
\item{x}{input data: vector or data frame with event times in seconds since tagon.}

\item{start_x}{name (no quotes needed) of variable in \code{x} that gives event times. Not needed if \code{x} is a vector or has only one column.}

\item{rl_data}{dataset with RL data (output of \code{\link[FBtagtools]{extract_rls}})}
}
\value{
A data.frame with the data and additional columns exposure_EventID (the ID number of the exposure event(s) during which this timestamp falls) and event_type (the type(s) of exposure events that were going on at this timestamp)
}
\description{
Add time (in seconds since tagon and perhaps as datetimes in UTC and/or local timezone) to a tibble or data-frame with regularly sampled data
}
