% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_all_clicks.R
\name{combine_all_clicks}
\alias{combine_all_clicks}
\title{Dive-level acoustic and tag data, for one or more tags}
\usage{
combine_all_clicks(
  tag_id = zc_smrt_tag_list,
  nc_path = getwd(),
  ae_path = getwd(),
  rl_file,
  ping_log_file,
  email,
  save_csv = TRUE,
  click_csv_name = "combined_zc_clicks.csv",
  buzz_csv_name = "combined_zc_buzzes.csv"
)
}
\arguments{
\item{tag_id}{Character string or vector with tag IDs (without "-cal.nc"). Default: all SMRT/Lander2 ziphius tags.}

\item{nc_path}{Directory (quoted string) where .nc files are stored. Can be one string, or a list the same length as tag_ids. Note: to download latest versions from Google drive, try function: \code{\link[FBtagtools]{download_drive_nc}}. Default: current working directory. Note: use "/" and not "\" to avoid headaches.}

\item{ae_path}{Directory (quoted string) where text files with info about acoustic events are stored. If needed, you can use \code{\link[FBtagtools]{download_drive_acoustic_events}} to get these. Default is the current working directory.}

\item{rl_file}{name (with path, if needed) of locally-stored .csv file with raw RL data. If not provided, the default is to use \code{\link[FBtagtools]{download_drive_rls}} to obtain it from the FB Google Drive.}

\item{ping_log_file}{name (with path, if needed) of locally-stored .csv file with raw RL data. If not provided, the default is to use \code{\link[FBtagtools]{download_drive_rls}} to obtain it from the FB Google Drive.}

\item{email}{Email address (required for FB Google Drive authentication; optional if \code{rl_file} is provided). You may also be asked to sign in or verify your Google identity as this function runs.}

\item{save_csv}{Logical; whether or not to save a csv file with the results. Default: TRUE (and will overwrite any existing file)}

\item{click_csv_name}{File name (with path, if desired) in which to save results in csv format. Default is combined_zc_clicks.csv.}

\item{buzz_csv_name}{File name for buzz file. Default combined_zc_buzzes.csv}
}
\value{
A named list of two tibbles, one for clicks and one for buzzes
}
\description{
Summarize SMRT (and/or Lander2) tag data on all clicks; add data on dive depth and sound exposure.
}
