% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ping_logs.R
\name{combine_ping_logs}
\alias{combine_ping_logs}
\title{Combine ping logs (acoustic audit data on anthropogenic signals) for one or more tags}
\usage{
combine_ping_logs(
  tag_id = zc_smrt_tag_list$tag_id,
  ping_log_path,
  nc_path = ping_log_path
)
}
\arguments{
\item{tag_id}{Character string or vector with tag IDs (without "-cal.nc"). Default: all SMRT ziphius tags.}

\item{ping_log_path}{Directory (quoted string) where ping log files are stored. Can be one string, or a list the same length as tag_ids.}

\item{nc_path}{Directory (quoted string) where nc files are kept (for fetching tag recording start times). Default: same as ping log path}
}
\value{
A data.frame() with information about timing of anthropogenic sounds on all tagouts in tag_id
}
\description{
Summarize SMRT (and/or Lander2) tag data from .nc files for each foraging dive cycle. There are data at two, nested time-scales: the dive cycle (a foraging dive and all the dives that follow it) and, within that, time-periods of set length in which other metrics are averaged. This function does coarse scale only. The data are intended for use in a state-switching model.
}
