% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_acoustic_summary.R
\name{dive_acoustic_summary}
\alias{dive_acoustic_summary}
\title{Dive-level acoustic and tag data, for one or more tags}
\usage{
dive_acoustic_summary(
  tag_id = zc_smrt_tag_list,
  nc_path = getwd(),
  ae_path = getwd(),
  bathy_path,
  ETOP01_bathy = FALSE,
  rl_file = c("RLs_3obank.csv", "RLs_3obank_2022.csv"),
  ping_log_file = c("qPing_log_corr_times_master",
    "Zica-20220112-195994_Individual_MFA_Pings.csv",
    "Zica-20211113-195993_Individual_MFA_Pings.csv",
    "Zica-20211112-94819_Individual_MFA_Pings.csv"),
  email,
  save_csv = TRUE,
  csv_name = "dive_acoustic_summary.csv"
)
}
\arguments{
\item{tag_id}{Character string or vector with tag IDs (without "-cal.nc"). Default: all SMRT ziphius tags.}

\item{nc_path}{Directory (quoted string) where .nc files are stored. Can be one string, or a list the same length as tag_ids. Note: to download latest versions from Google drive, try function: \code{\link[FBtagtools]{download_drive_nc}}. Default: current working directory. Note: use "/" and not "\" to avoid headaches.}

\item{ae_path}{Directory (quoted string) where text files with info about acoustic events are stored. If needed, you can use \code{\link[FBtagtools]{download_drive_acoustic_events}} to get these. Default is the current working directory.}

\item{bathy_path}{A directory path to the folder containing all bathymetry data. Use \code{\link[FBtagtools]{download_bathy}} if you don't have this data already. If not provided, the bathymetry data will not be included in the output dataset.}

\item{ETOP01_bathy}{whether to try to fill-in bathymetry info for locations outside NEPAC dataset bounds using NOAA online ETOP01 database. Default is FALSE. This is slow and no animals have gone out-of-bounds yet as of July 2021.}

\item{rl_file}{name(s) (with path, if needed) of locally-stored .csv file(s) with raw RL data. If not provided, the default is to use \code{\link[FBtagtools]{download_drive_rls}} to obtain it from the FB Google Drive.}

\item{ping_log_file}{name(s) (with path, if needed) of locally-stored .csv file(s) with ping log (acoustic audit) data. If not provided, the default is to use \code{\link[FBtagtools]{download_drive_rls}} to obtain it from the FB Google Drive.}

\item{email}{Email address (required for FB Google Drive authentication; optional if \code{rl_file} is provided). You may also be asked to sign in or verify your Google identity as this function runs.}

\item{save_csv}{Logical; whether or not to save a csv file with the results. Default: TRUE (and will overwrite any existing file)}

\item{csv_name}{File name (with path, if desired) in which to save results in csv format. Default is dive_acoustic_summary.csv.}
}
\value{
A data.frame() with one row per dive, per whale
}
\description{
Summarize SMRT (and/or Lander2) tag data from .nc files for each foraging dive cycle. Note: Currently for only SMRT tags; function and help will be updated to allow inclusion of Lander2 data as well when possible.
}
\examples{
# Examples will go here
}
