% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_cycle_summary.R
\name{dive_cycle_summary}
\alias{dive_cycle_summary}
\title{Dive-cycle-level acoustic and tag data, for one or more tags}
\usage{
dive_cycle_summary(
  tag_id = zc_smrt_tag_list,
  nc_path = getwd(),
  ae_path = getwd(),
  bathy_path,
  ETOP01_bathy = FALSE,
  rl_file,
  ping_log_file,
  email,
  save_csv = TRUE,
  csv_name = "dive_cycle_summary.csv"
)
}
\arguments{
\item{tag_id}{Character string or vector with tag IDs (without "-cal.nc"). Default: all SMRT ziphius tags.}

\item{nc_path}{Directory (quoted string) where .nc files are stored. Can be one string, or a list the same length as tag_ids. Note: to download latest versions from Google drive, try function: \code{\link[FBtagtools]{download_drive_nc}}. Default: current working directory. Note: use "/" and not "\" to avoid headaches.}

\item{ae_path}{Directory (quoted string) where text files with info about acoustic events are stored. If needed, you can use \code{\link[FBtagtools]{download_drive_acoustic_events}} to get these. Default is the current working directory.}

\item{bathy_path}{A directory path to the folder containing all bathymetry data. Use \code{\link[FBtagtools]{download_bathy}} if you don't have this data already. If not provided, the bathymetry data will not be included in the output dataset.}

\item{ETOP01_bathy}{whether to try to fill-in bathymetry info for locations outside NEPAC dataset bounds using NOAA online ETOP01 database. Default is FALSE. This is slow and no animals have gone out-of-bounds yet as of July 2021.}

\item{rl_file}{name (with path, if needed) of locally-stored .csv file with raw RL data. If not provided, the default is to use \code{\link[FBtagtools]{download_drive_rls}} to obtain it from the FB Google Drive.}

\item{ping_log_file}{name (with path, if needed) of locally-stored .csv file with raw RL data. If not provided, the default is to use \code{\link[FBtagtools]{download_drive_rls}} to obtain it from the FB Google Drive.}

\item{email}{Email address (required for FB Google Drive authentication; optional if \code{rl_file} is provided). You may also be asked to sign in or verify your Google identity as this function runs.}

\item{save_csv}{Logical; whether or not to save a csv file with the results. Default: TRUE (and will overwrite any existing file)}

\item{csv_name}{File name (with path, if desired) in which to save results in csv format. Default is dive_acoustic_summary.csv.}

\item{rl_model_dir}{directory where files with results of RL modeling (for non-acoustic tags) are stored locally}

\item{dive_class_file}{name (with path, if needed) of locally-stored .csv file with information on machine-learning classification of dives as foraging or not. This should match perfectly with presence/absence of clicking from acoustic recordings, but is mainly for tags/dives without acoutics.}
}
\value{
A data.frame() with one row per dive, per whale
}
\description{
Summarize SMRT (and/or Lander2) tag data from .nc files for each foraging dive cycle. There are data at two, nested time-scales: the dive cycle (a foraging dive and all the dives that follow it) and, within that, time-periods of set length in which other metrics are averaged. The data are intended for use in a state-switching model.
}
