% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bathy.R
\name{download_bathy}
\alias{download_bathy}
\title{Download bathymetry data, from FB google drive}
\usage{
download_bathy(email, bathy_path = getwd(), overwrite = TRUE)
}
\arguments{
\item{email}{Email address (for FB Google Drive authentication). You may also be asked to sign in or verify your Google identity as this function runs.}

\item{bathy_path}{directory in which to save the zip file. Defaults to the current working directory.}

\item{overwrite}{Whether or not to overwrite existing zip file. Default: TRUE}
}
\value{
A zip file with bathymetry data will be downloaded. To use, you'll then need to unzip it.
}
\description{
Download bathymetry data, from FB google drive
}
\examples{
#examples not yet provided, sorry :(
}
