% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_drive_acoustic_events.R
\name{download_drive_acoustic_events}
\alias{download_drive_acoustic_events}
\title{Download spreadsheets with information about acoustic events from the FB Google Drive}
\usage{
download_drive_acoustic_events(
  tag_id = zc_smrt_tag_list,
  ae_path = getwd(),
  email,
  overwrite = TRUE,
  type = c("AllClicks", "Buzzes", "PostProcessedEvents")
)
}
\arguments{
\item{tag_id}{Character string or vector with tag IDs to download (without "-cal.nc" and without additional file path information). Default: all SMRT ziphius tags.}

\item{ae_path}{Directory (quoted string) where spreadsheet files should be stored locally. Can be one string, or a list the same length as tag_ids. Default: current working directory. Note: use "/" and not "\" to avoid headaches. Final "/" not necessary.}

\item{email}{Email address (for FB Google Drive authentication). You may also be asked to sign in or verify your Google identity as this function runs.}

\item{overwrite}{Logical; whether to overwrite existing files. Default is TRUE.}

\item{type}{The types of acoustic events files to find and download. Options are one or more of: 'AllClicks', 'Buzzes', 'PostProcessedEvents'. Default is all three.}
}
\value{
Returns a "dribble" with information about the files and their location on google drive. (The function also downloads the requested files, of course.)
}
\description{
Default is to download data related to all Ziphius SMRT tag datasets (or you can specify the tag(s) for which you want to download data). Data are downloaded from the FB project shared Google Drive.
}
\examples{
download_drive_nc(email = "sld33@calvin.edu") # (this only works if you know sld33's password...)
}
