% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rls.R
\name{extract_rls}
\alias{extract_rls}
\title{extract RL data}
\usage{
extract_rls(
  rl_file = NULL,
  ping_log_file = NULL,
  email,
  save_output = TRUE,
  path = getwd(),
  out_file = NULL,
  signal = "MFAS",
  overwrite = TRUE
)
}
\arguments{
\item{rl_file}{name (with path, if needed) of locally-stored .csv file with raw RL data. If not provided, the default is to use \code{\link[FBtagtools]{download_drive_rls}} to obtain it from the FB Google Drive.}

\item{ping_log_file}{name (with path, if needed) of locally-stored .csv file with raw RL data. If not provided, the default is to use \code{\link[FBtagtools]{download_drive_rls}} to obtain it from the FB Google Drive.}

\item{email}{Email address (required for FB Google Drive authentication; optional if \code{rl_file} is provided). You may also be asked to sign in or verify your Google identity as this function runs.}

\item{save_output}{Logical; whether or not to save results in a .csv file. Default is TRUE.}

\item{path}{Quoted string with the path to the directory where you want to save the output file. Defaults to the current working directory. Final "/" not needed. Use "/" rather than "\" to avoid possible headaches.}

\item{out_file}{Name (quoted string) for output .csv file with results. Optional; default file name is constructed based on signal type and frequency band requested.}

\item{signal}{Quoted string (or vector of them) indicating which signal types to return RLs for. Options are one or more of: 'MFAS', 'Echosounder', 'Explosive'. Default: 'MFAS'. Note: CAS is currently marked as MFAS type. You can isolate CAS pings and events by getting MFAS RLs and then keeping only pings with duration 20 seconds or longer.}

\item{overwrite}{Whether or not to overwrite an existing output file. Logical. Default: TRUE.}
}
\value{
Returns a data.frame with RLs (one row per ping). If \code{save_output} is true, also saves a csv file in directory \code{path} with filename \code{out_file.csv} with the results.
}
\description{
This function processes the "raw" format of RL data, allowing users to obtain RLs of desired types and in selected frequency bands. If you need direct access to the (more complicated) raw data, see \code{\link[FBtagtools]{download_drive_rls}}.
}
\examples{
extract_rls(email = "sld33@calvin.edu") # (this only works if you know sld33's password...)
}
