% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_wav_times.R
\name{extract_wav_times}
\alias{extract_wav_times}
\title{extract timing info for SMRT wav files}
\usage{
extract_wav_times(tag_id = zc_smrt_tag_list, nc_path = getwd())
}
\arguments{
\item{tag_id}{Deployment IDs for the tag(s) for which you want to pull information. Defaults to the Ziphius deployments in \code{zc_smrt_tag_list}.}

\item{nc_path}{Quoted string with the path to the directory where nc files are stored. Defaults to current working directory.}
}
\value{
Returns a data.frame with information about each wav file for each deployment of interest. If there are multiple rows per wav file, that is indicative of a (filled or un-filled) gap; the first row for a given deployment gives the file start time.
}
\description{
This function pulls data on start times (and other metadata) about acoustic wav files recorded by SMRT tags. It assumes that you already have downloaded the nc files for the deployments of interest to you; if not, see \code{\link[FBtagtools]{download_drive_nc}}
}
\examples{
extract_rls(email = "sld33@calvin.edu") # (this only works if you know sld33's password...)
}
