% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tagon_time.R
\name{get_tagon_time}
\alias{get_tagon_time}
\title{Get tagon time from an nc file. Returns one datetime.}
\usage{
get_tagon_time(tag_id, nc_path)
}
\arguments{
\item{tag_id}{Character string or vector with tag IDs (without "-cal.nc"). Default: all SMRT ziphius tags.}

\item{nc_path}{Directory (quoted string) where nc files are kept (for fetching tag recording start times). Default: same as ping log path}
}
\value{
a list of datetimes (when the tag record(s) started)
}
\description{
Read in data from .nc file to obtain the tag recording start time
}
