% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_join.R
\name{interval_join}
\alias{interval_join}
\title{Join data at different time-scales}
\usage{
interval_join(
  x,
  y,
  start_x,
  start_y,
  end_x = start_x,
  end_y,
  suffix = c("", ".new"),
  ...,
  keep = FALSE
)
}
\arguments{
\item{x}{interval data frame or tibble (data with one row per dive, dive-cycle, day, etc.)  Should have columns that provide start and end times of each interval; these can be date-times or any numeric time indicator}

\item{y}{data frame at finer time-scale with information to summarize and pull into \code{x}}

\item{start_x}{name of variable with interval start times in \code{x}}

\item{start_y}{name of variable in \code{y} that contains sample times or event-start times}

\item{end_x}{name of variable in \code{x} that contains interval end times (defaults to \code{start_x} if not given)}

\item{end_y}{name of variable in \code{y} that contains event or sample end times (defaults to \code{start_y} if not given)}

\item{suffix}{(Passed to \code{\link[dplyr]{left_join}}; most users can ignore) If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.}

\item{...}{Additional arguments to pass to \code{\link[dplyr]{left_join}}}

\item{keep}{(Passed to \code{\link[dplyr]{left_join}}; most users can ignore and keep the default value, FALSE) Should the join keys from both x and y be preserved in the output?}
}
\value{
A data.frame like the input interval dataset \code{x}, but with additional columns for the new summarized variables
}
\description{
Given one dataset where each row represents an interval at coarser scale (for example, one row per dive with dive start/end times), pull in summary information about finer-time-scale data. For example, add the max RL during each dive, or the median MSA during each dive.
}
\examples{
Examples will go here
}
