% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_elev.R
\name{solar_elev}
\alias{solar_elev}
\title{Add time columns to time-series}
\usage{
solar_elev(time, lat, long, time_zone = "America/Los_Angeles")
}
\arguments{
\item{lat}{latitude}

\item{long}{longitude}

\item{time_zone}{Time zone for local time; defaults to 'America/Los_Angeles'}

\item{utc_time}{UTC time}
}
\description{
Add time (in seconds since tagon and perhaps as datetimes in UTC and/or local timezone) to a tibble or data-frame with regularly sampled data
}
\examples{
Examples will go here
}
