% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_stage.R
\name{solar_stage}
\alias{solar_stage}
\title{Add time columns to time-series}
\usage{
solar_stage(utc_time, lat, long, time_zone = "America/Los_Angeles")
}
\arguments{
\item{utc_time}{UTC time}

\item{lat}{latitude}

\item{long}{longitude}

\item{time_zone}{Time zone for local time; defaults to 'America/Los_Angeles'}
}
\value{
solar stage (categorical): "Day", "Night", "Dusk" or "Dawn"
}
\description{
Add time (in seconds since tagon and perhaps as datetimes in UTC and/or local timezone) to a tibble or data-frame with regularly sampled data
}
\examples{
# Examples will go here
}
