% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc_to_local.R
\name{utc_to_local}
\alias{utc_to_local}
\title{Add local time column to data frame with UTC time}
\usage{
utc_to_local(x, utc_var, tz = "America/Los_Angeles")
}
\arguments{
\item{x}{input data must contain a column that gives the UTC times to be converted to local}

\item{utc_var}{name of variable (in \code{x}) that gives the UTC times}

\item{tz}{desired time zone name -- for example the default, 'America/Los_Angeles'}
}
\value{
data.frame \code{x} with one added column called local_time
}
\description{
Given an input data frame that already has UTC datetimes, add another datetime column giving the local time
}
\examples{
# Examples will go here
}
