% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_acoustic_summary.R
\name{dive_acoustic_summary}
\alias{dive_acoustic_summary}
\title{Dive-level acoustic and tag data, for one or more tags}
\usage{
dive_acoustic_summary(tag_id = zc_smrt_tag_list, path = "")
}
\arguments{
\item{tag_id}{Character string or vector with tag IDs (without "-cal.nc"). Default: all SMRT ziphius tags.}

\item{path}{Directory (quoted string) where .nc files are stored. Can be one string, or a list the same length as tag_ids. Note: to download latest versions from Google drive, try function: \code{\link[FBtagtools]{download_drive_nc}}. Default: current working directory. Note: use "/" and not "\" to avoid headaches.}
}
\description{
Summarize SMRT (and/or Lander2) tag data from .nc files for each foraging dive cycle. Note: Currently for only SMRT tags; function and help will be updated to allow inclusion of Lander2 data as well when possible.
}
\examples{
Examples will go here
}
