% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_rls.R
\name{sum_rls}
\alias{sum_rls}
\title{Sum RLs in dB}
\usage{
sum_rls(rls)
}
\arguments{
\item{rls}{numeric vector with RLs in dB}
}
\value{
sum of the received levels in \code{rls}
}
\description{
Add up RLs in several frequency bands. Used by \code{\link[FBtagtools]{extract_rls}}.
}
\examples{
# coming soon
}
